+++ scripts/macros/animations.mjs
@@ -9,3 +9,4 @@ export {cunningActionDash} from './animations/cunningActionDash.mjs';
 export {cunningActionDisengage} from './animations/cunningActionDisengage.mjs';
 export {rageDefault, rageLightning, rageSaiyan} from './animations/rage.mjs';
 export {instinctivePounce} from './animations/instinctivePounce.mjs';
+export {smite} from './animations/smite.mjs';